Telephony: Send user confirmation instead of Terminal Response
On certain proactive commands like  Setup Call/Open Channel,
a confirmation dialog is displayed to the user to confirm or reject the
call. If the user hits the back key to dismiss the dialog or the user
hits 'No', consider that as the user rejecting the call, instead
of sending a Terminal Response.

CRs-fixed: 411548
(cherry picked from commit 03660f49cde06176c46fdd5e94e0ee844e70d714)

Change-Id: I2da7ab45ed027ba867e1a0d55216863ed139bcf1