Telephony: Send user confirmation instead of Terminal Response
On certain proactive commands like  Setup Call/Open Channel,
a confirmation dialog is displayed to the user to confirm or reject the
call. If the user hits the back key to dismiss the dialog or the user
hits 'No', consider that as the user rejecting the call, instead
of sending a Terminal Response.

Change-Id: I489c790d398ff8347636756a6736594e46e12ccc
CRs-fixed: 411548