Bluetooth: Avoid deadlock in aquiring bluetooth service in rare scenarios
While doing quick BT On/Off, at times both On/Off seuence trying
to get the access to bluetoothservice object leading to this issue,
In other case while trying to update the SDP records, if event loop
get the response from the controller very fast or with some reason
thread stuck in updating SDP records this issue is observed.
Removed the syncronization for updating the uuid info to resolve
this issue.

Change-Id: I7c32e80ff90c033c91785961a9927883282eea9c
CRs-fixed: 452659