DUT display "Select all/cut/copy" menu when add a new contact.
When we edit in landscape mode, it usually uses
the ExtractEditLayout to handle the edit action. In this view,
when we long touch the words, and click the edit button, it will
show us a dialog which do the same thing as Editor(which allow
us to do copy, move, paste, etc.). However, the dialog hasn't
been dismissed when the input window is hidden. So it will be
shown again when we start a new edit action.

Dismiss the dialog in ExtractActionMode's finish() method.

CRs-Fixed: 412457

Change-Id: I35401a5da055b5b7136d4ba2177add59c793ffdb