The contact number doesn't display completely when reply SMS.
There are two ways to show the ellipsize when the text
width is larger then the content width.One is showing a dot,
the other is showing three dot. It decides by whether the
textView has the "android:ellipsize" property or not.

Let TextView show three dot with "android:ellipsize" property.

Change-Id: Ica1c2417c2b356f5b5950aabd7e56254331f9ca0
CRs-Fixed: BZ405