frameworks/base: Fix for noise on headset during an MT call with camcording
Issue: Noise is heard on wired headset when there's an incoming
call during video recording.

Description: When there's an incmoing call during video
recording with wired headset, shutter sound and ringtone
are to be played on combo devices: speaker+headset. However,
shutter sound (which is a system sound) was routed to headset
and ringtone is routed to speaker+headset, because of which
switchDevice happens twice and incurs delay thereby resulting
in initial noise.

Force enabled the camera shutter sounds in config.xml to play
them on speaker.

CRs-Fixed: 475424

Change-Id: I3237c77c4385d77743050379cb036c8949fb6e35