Android Policy: Make Volume Keys wake up capable
Currently Android does not wake up the system if volume keys are pressed.
This is to enable the user to change volume while keeping
the screen is off. In AM335X device volume keys are the only
keys to wake up domain(GPIO Bank 0). So changing this default behavior
to bring the Unlock Screen when pressing Volume Keys.

Signed-off-by: Pankaj Bharadiya <pankaj.bharadiya@ti.com>