Reduce total number of background apps for low memory devices
Due to memory constraints, we anyway cannot fit the default number
of BG apps. So instead of killing background apps to
free up memory when launching, start with an already
reduced background set of apps.
Another side-effect is that launch latencies also improve.

Note: This is for low memory devices. (~512MB RAM or less)

Change-Id: I5d42bf58a77cd4ce460ba655a1365031b144aaf1