The contact number doesn't display completely when reply SMS.
There are two ways to show the ellipsize when the text
width is larger then the content width.One is showing a dot,
the other is showing three dot. It decides by whether the
textView has the "android:ellipsize" property or not.

Let TextView show three dot with "android:ellipsize" property.

CRs-Fixed: BZ405

Change-Id: I300a423408120d7ff1a2594ea1f840924b79496d