InputMethod: dismiss the popupwindow and launch the camera
As you press the edit button to edit message content,
the popwindow pops up.And then you press the H/W camera
key with long press to start the camera,
the popwindow covers the camera interface.
So the camera application can not launch successfully

Capture the long press event of camera
via onkey of MenuPopupHelper.When you Capture the event,
you can dismiss the popwindow and launch the camera.

CRs-Fixed: 510614

Change-Id: Ia77531e59578c55c236390a70888f89f19526cdb