Keyguard: Distinguish the TextView id of pinEntry
Both KeyguardSimPinView and KeyguardPinView have the
same id of pinEntry for their TextView of the password.
When one of them is created in the mSecurityViewContainer,
the other one will get the wrong TextView of the same
id pinEntry.

Change the id pinEntry in KeyguardSimPinView to simPinEntry.
Add the new id simPinEntry to symbols.xml.

CRs-Fixed: 522675

Change-Id: Ie6627448820d8c526c298c0c14e032c845bc9d88