Webkit: reset media player when webview goes background
This patch resets the mediaplayer in webview when it
goes to background. Also prepares the player for play
as soon Webview comes to foreground.

This is required as the mediaplayer instance is not shared across
webviews and there are times decoder is not able to allocate buffers
unless the reset is done on instances in the background.

Change-Id: I8cc10c33a2478ed682536c602f876255e4261f7e