Common: Remove a dead lock in power manager
The two threads (thread to pause a activity and thread to try to wake up
when usb plugged or unplugged) will wait for two locks (AcivityManagerService
self, mLock in PowerManagerService). and it will cause a dead lock when these
two threads are started at the same time, so remove AcivityManagerService lock
in the thread to try to wake up when usb plugged or unplugged.
Remove the function to get the state of fast boot by getting running proceses
which will wait for AcivityManagerService as a lock, and replace it with a
property which stored the state of fast boot.

CRs-Fixed: 537187

Change-Id: I586852995c10e0f0fbc278bd3adf8d6bb571703d