add dedicated alarm type for poweroff alarm
Now we use type of RTC_WAKEUP for the poweroff alarm.
Actually lots of other applications also use this type.
But these applications never expect the device will
power on after it has been shut down. So we need to
add dedicated alarm type for the Deskclock or similar
applications.

CR-fixed: 529674

Change-Id: I8cc666c9482ef5c624e13277ac6855b6ea16b290