Magnification: Fix screen freeze when activity resume with rotation change
When magnification is enable, if the rotation of activity which will goto
is changed, the activityManager will crash.

Cancel the magnification animation must run in a thread which has looper.
But the main thread isn't have looper.

Move the call of mMagnifedViewport.onRotationChangedLocked() to the handler
which created with looper.

CRs-fixed: 541799

Change-Id: I2a78f24f73d5a9b8693e557d18e2fb01fcaec8fa