StorageFormatter: Add support to multi storages
When multi storages are broken at the same time.ExternalMediaFormatActivity
don't support to erase all the storages.Due to there is only one
storagenotification with this FORMAT intent.
Even if there are two seperate StorageNotifications for multi storages,it
still doesn't work.Because of the principle of PedningIntent.
Only one parameter can be put into these PendingIntent.
The ExternalMediaFormatActivity will format the same the storage.
The other one still is  blank or broken, and can't be formated.

This change add a new formatactivity named PhoneMediaFormatActivity, actually
It's a copy of ExternalMediaFormatActivity.So there will be two
FormatActivity for two PendingIntent.

Change-Id: I8aceb6521ceb97c0c0546b9dc957d84e1da6baf3