webkit: avoid unnecessary seeking in html5 video
Without this patch, MediaPlayer.seekTo() is called in resuming.
For live streaming, seeking will cause error and cannot continue to
play. For other video, seeking may cause video backward for a short
period.

Check before seeking. If the new position is the same as the old, drop
the seeking.

Change-Id: If36f0d768711e2c042549c12b8a5bfacac487c73
CRs-Fixed: 543154