Settings: Fix "lockscreen mode can not be saved after removing battery"
LockSettingsService.writeFile() uses "rw" flag but no sync flag("s").
So, the password is not written to the physical device immediately.
When users change lockscreen mode and remove the battery directly in a
short time, the password will can not be saved.

Since the lockscreen password is an important user data, we should add
a sync flag to save it to the physical device immediately.

CRs-Fixed: 541772

Change-Id: Iab350aadb3058afbc545c48f82d6413126e18ed5