Bluetooth: Handle Incoming connection when unpairing has started
This patch handles Incoming HID conenction when unpairing has started in
upper layers. With this patch if there is an incoming HID connection from
device for which unpairing has started, connection request is rejected
instead of being ignored, so that BlueZ user space can update its
pending authorization flags. without this patch it was observed that it
was not possible to accept incoming connections until device reboot
because of BlueZ state going bad.

Change-Id: Iaa17a375df6dbc798fca44abc7ceeea1f1743c66
CRs-Fixed: 542825