base: App crash when suggestion's end more than textview's length
App crashed by catch IndexOutOfBoundsException which raised
by check span range when select suggested string.The root
cause is that the editor will truncate the suggested string
which length is more than it specified, but it also set span
use the suggestion's original length.

Don't setspan if the suggested string is truncated.

Change-Id: Ib405564fd99037601879d240bbf8b088c2543230