Settings: Fix 'EAP method'droplist can't be chosen after unlock DUT.
The current focus has changed to the "Add network" dialog window
below after unlock DUT, and this will cause the droplist window
loses focus and can not be chosen.

According to the window management service mechanism, the
visibility of the window changed may cause window focus change.
In the performShowLocked method, when show child window, the
state of the attached  parent window must be invisible to
visible, so mAttachedHidden changes to "false". If the stack
layer of child window which we are showing on the top of the
current focus, we should set mService.mFocusMayChange to "true",
so that call the updateFocusedWindowLocked() method to
re-evaluate focus and make sure it is correct.

CRs-fixed: 537132

Change-Id: Iebd1e6dcb633aaf100659443bcb3fd8654e87e02