Gallery2: fix slotview and action bar overlapped after rotation
The height of action bar is different in landscape and portrait
mode, it won't be updated until ActivityManagerService invoke
setNewConfirguration in WindowManagerService. But the slotview
will be redrawn before action bar, slotview will get the old
action bar height to calculate the height and width of thumbnail,
it get a greater value, then slotview and action bar overlapped.

Analysis updateConfigurationLocked in ActivityManagerService,
we can see that the onConfigurationChanged of AbsActionView
will be invoked before setNewConfirguration, we add to update
the action bar here, then the slotview can get the correct
height value of action bar.

CRs-Fixed: 562745

Change-Id: Id9c17ae618d72b44849f81146a2873444cdb467a