Revert "Settings: Disable Mobile data after tap OK on network over limit dialog"
This reverts commit db72c4cb4361830404eb2126daa750b58a91d9de

comment:
Mobile data will be disabled and apn will be deleted after tap OK on
network over limit dialog for CR524765,this is not appropriate.Because
the policy of limiting data usage just disabled mobile data for a while
but didn't want to delete the spn completly when we reached the
specified data usage limit,and enabled the mobile data after tap
"Re-enable data" button.

If we disable the mobile data through delete the spn,we can't enable the
mobile data by snooze the policy for CR548415.

So CR524765 violate the intention of the designer,we should revert its't
commit.

CRs-fixed: 548415

Change-Id: Iceae331b42859a9f1f781a6c980a72f27c1a2092