SystemUI: Fix System UI force close at SSSS and CT mode
SignalClusterView will be replaced by CTSignalClusterView in status bar,
When hide the statusbar, it will get SignalClusterView to start animation
but it will get null at this time. So the app will force close
when use that.

When replace SignalClusterView to CTSignalClusterView, set
CTSignalClusterView id to orign SignalClusterView id.

Change-Id: I3e0b88fde8f35b3fc925b550d27f187d7f416d4a
CRs-fixed: 568836