Browser: Web page is unresponsive after changing language
When current web page pops up a confirm dialog,
sendMessageToUiThreadSync() in CallbackProxy will be called to
pause the WebCoreThreadWatchdog and then call wait() to pause
current thread for waiting user's option. When back from Settings,
web page will be onResume(), and also resume the watchdog. But the
web page is still in waiting state, thus the watchdog will pop up
a unresponsive dialog after 30s' timeout.

We need lock the watchdog when webview call wait() in
sendMessageToUiThreadSync() and unlock the watchdog when webview
have been synchronized with user's operation.

CRs-Fixed: 539514

Change-Id: I277f3fb8ac8401aca9e822c9dc0b3ba50308b0ff