Browser: failed to show input character completely for finding-on-page
In webview_find.xml, EditText for inputting and TextView for
showing matches string share the same space on action bar.
When system language is Chinese, the string shown in TextView
is too long and then the width of EditText become too short to
show a character completely.

Delete the blank spaces between numbers and Chinese characters to
compress TextView's space.

CRs-Fixed: 539205

Change-Id: I19cca7f2126ffe1fd9325facf4ccbfd457f41534