Mms: fix web url can't be recognised in sms conversation
Google think the web url should be end with a letter.So the
web url's regular expression will match the last letter of the url.
The string will not be regarded as a web url if the end word
of the string isn't a English letter.

Cut the match rules which will match the last letter of the string.

CRs-Fixed: 579781

Change-Id: Ie222bf1d2ac969e397fd962a83c8a6b2eab5ebf7