Mms: make web_url match can pass cts test
For testAddLinks1 cts will add a "," in the end the url.But google
will cut the "," when it match the url.So if we match the "," in
the regex matcher will cause "," will be regarded as a part of the
web url.

Cut the "," matches in the web url regex matcher to fix cts test failed
issue.

CRs-Fixed: 549791

Change-Id: I693e708db4bda937c3ceab364d33da97b9164f66