SystemUI: Fix operator name showing "NULL"
As getting network operator name returns "null" string which comes from
ril when having no service, device will show "NULL" in notification bar.

Add a "null" string check for it. If it's "null", we will set it to "",
so that device will show default network name in notification bar.

CRs-Fixed: 583964

Change-Id: I28293360e379e4e19bf61e43ff61b566ba16aedb