WiFi: Enable roaming to the SSID with GBK Character set.
There is a new change to use bssid to distinguish GBK and UTF
charset ssid passed down from framework. This is to support APs
with GBK Chinese character.But it will cause the roaming issue
as bssid will be different for same SSID. In supplicant layer,
it will first check BSSID,if it is different, the roaming will
not happen.So here always store the original ssid and use it
for all kinds of WiFi operations.If the charset is GBK,then it
will convert to UTF8 for upper layer to use it and also keep the
mapping for GBK SSID vs UTF8 SSID.When UTF-8 SSID is passed down
from upper layer, it will checking the mapping.If there is a GBK
SSID associated, then will pass down the GBK SSID to lower layer.

Change-Id: I59002052aad79ff3e2d99419345b6a1e8b493a38
CRs-Fixed: 580358