WiFi: Fix the roaming issue.
There is a new change to use bssid to distinguish GBK and UTF
charset ssid passed down from framework. This is to support APs
with GBK Chinese character.But it will cause the roaming issue
as bssid will be different for same SSID. In supplicant layer,
it will first check BSSID,if it is different, the roaming will
not happen.So here always store the original ssid and use it
for all kinds of WiFi operations.

Change-Id: I59002052aad79ff3e2d99419345b6a1e8b493a38
CRs-Fixed: 580358