Keyguard: Fix click del button once,two charactors deleted.
When click del button once,the function of dispatchKeyEvent
is executed twice. One is in the class of KeyguardViewBase,
the other is in the class of KeyguardViewManager.

When the Key is not the media key,it is not nessacery to execute
the dispatchKeyEvent in the KeyguardViewBase.So add a condition
to make sure when the del button was clicked,only dispatchKeyEvent
in the KeyguardViewManager was excuted.

CRs-Fixed: 589194

Change-Id: If405e52e0fad7a427b6d47fddabd4219329ff1d7