System UI: Fix back key invalid when inputMethod process being killed
InputMethodManager instance is not null, so imm is not null, but
InputMethodManager.mCurMethod will be set to null when disable or kill
IME. It's normal behavior, and this is not synchronization issue. Google
return FINISH_NOT_HANDLED but then the next input stage doesn't have the
chance to handle this event, and at last, after going through all input
stages, this event will be dropped. So change FINISH_NOT_HANDLED to
FORWARD is to make sure this keyevent can have chance to be processed by
the next input stage when ime is disable or been killed(ime can easily
be killed on low ram device), it should have little side effect.

CRs-Fixed: 588608
Change-Id: Id45ac4b5b13fe6fec1f9548a2479811b00b56041