Revert "Fix for Concurrency issue reported in  525529"
The fix was introduced to avoid a rare race condition
in ActivityManager.With the fix mServices.serviceTimeout()
is called while holding ActivityManagerService.this lock.
This causes appNotResponding(), which processes ANRs and dumps
extensive trace data being called with AM locked.And it
leaves AM not responding to requests from other processes
during this time and may introduce ANRs in applications.

This reverts commit cb79b6323aa455bc235eb04f7574e5d4da912c12.

CRs-Fixed : 585504

Change-Id: I982b45d8d7d441d076274e999e3894a248f6ff7b