Webkit: fix cannot show "no matches" when match count is zero
Google's previous refactor for FindActionModeCallback results
in this issue, this makes WebView call FindActionModeCallback::
onFindResultReceived when WebView wants to update match count,
in onFindResultReceived function, it considers this find result
as empty if match count equals zero and so cause not showing
"no matches" string in TextView.

We correct the parameter for empty find with false value in
onFindResultReceived function.

Change-Id: I0343af65829a00ac68a782b61bd053f28a81700b
CRs-Fixed: 599239