Report actual error code for multisim PIN/PUK operations.
The current APIs that are used to verify the PIN and PUK only
convey whether the operation succeeded or failed.

Add 2 new APIs that report the actual error code
'PASSWORD_INCORRECT' and 'GENRERIC_FAILURE' in case of failure.

Avoid overwriting the security message which displays the number
of attempts remaining with default reset message.

Change-Id: I04109247a4ffe9d0908d2282ce8dae6303f65b22
CRs-Fixed: 655989