StatusBar: Fix only one carrier text shown under DSDS
When config_display_rat is enabled, RAT info will be appended
to the reported plmn or spn, which only happens when data or
voice is in service by design. When the SIM card is ready and
data/voice is out of service, plmn or spn will be converted to
empty string.

Add code to fix it: when data and voice is out of service, do
nothing to spn or plmn, rather than converting them to empty
string.

Change-Id: Icf232011f833e3b590f49d401c2d49cf88d45fe9
CRs-Fixed: 693743