Use TelecomManager to call TelephonyManager#getVoicemailNumber (1/2).
Since we don't want the phone app to call TelephonyManager methods
directly, create a method in TelecomManager to mimic the behavior. The
reason we don't return the voicemail number directly is because once
TelecomManager becomes public, we don't want any old app to be able to
get the SIM's voicemail number.

Bug: 18233678

Change-Id: I620af409788aea7b58169b72a563baeed8115080