SoFTAP: fix multi tethering icons display
By default, when tethering is enabled, no matter single or
multiple tethering case, there will be only one icon display
in notification area.

But with CMCC SAP feature, in multi tethering case, multi
tethering icons will be displayed in notification area, and
will not be cleared even tethering is disabled.

This change is to restore the tethering icons display
behavior to default.

Change-Id: I59c48a43a3b706109b14f07019990c10a925471a
CRs-Fixed: 770921