NotificationLED: enable notification led blink when screen is off
By default, notification led will not blink if the device screen
is off, but actually, 'blinking when screen is off' is a very
important feature for end user, there are a lot of high-end smart
phones which integrate RGB led and need the leds blink even when
screen is off, for example: when user receives a missed call, or
a missed sms, the RGB led will blink when screen is off.

To support this, call updateNotificationPulse() when screen is off.

Change-Id: Ib35353514c2af868e37098e5b4e2cd0c8b8fae78