KeyGuard: Fix incorrect SIM 1 messages wrongly displayed for SIM 2.
Failing to enter the correct PIN for SIM1 in the available attempts
results in having to input the PUK code.Upon unlocking the SIM 1
with the correct PUK code, SIM 2 is supposed to display the SIM PIN
message but instead it is wrongly displaying the PUK message.
Fix includes to clear the previous states during reset in
KeyGuardSimPinView and show the default PIN message for SIM 2.

CRs-Fixed: 774032.

Also, currently SIM 1 gets disabled after failing to input the correct
PIN and PUK code.When user enters wrong PIN for SIM 2 and exceeds the
available PIN attempts, a message to input PUK code is supposed to
display, instead it is now wrongly displaying the SIM disabled message.
Fix includes to clear the  previous states during reset in
KeyGuardSimPukView and show the default PUK message for SIM 2

Change-Id: Ie5c90a37f3252ecae35ee0ee27cf772837f9a569
CRs-Fixed: 778839.