Added 'wipe' option to 'encryptStorage' MountService API
The 'wipe' option in 'encryptStorage' API is used to create a new ext4
file system on the userdata partition, instead of the existing one,
and encrypt it.
This as opposed to the current default option 'inplace',
that encrypts the existing file system as is.
The 'wipe' option is already supported in the underlying Cryptfs
implementation -'enable_crypto' and this change exposes this option to
the MountService API.

Also in this change, new values that can be returned by
'getEncryptionState' API are declared.
These values reflects the state of the MDTP activation, together
with the general encryption state, in case that MDTP is activated.
- ENCRYPTION_STATE_OK_MDTP_ACTIVATED - means that the crypto state is ok,
and MDTP is activated.
- ENCRYPTION_STATE_ERROR_MDTP_ACTIVATED - means that the crypto state is
bad, and MDTP is activated.

Change-Id: Ide628a8cf6499bc2216b08c22479a37133bebb03