ActivityManager: use Process.killProcessGroup to kill forked processes
Place newly-created processes in a process group, and use
killProcessGroup to kill any forked processes

Add Process.killProcessGroup to interface between ActivityManager and
libprocessgroup.

Try to catch any cases where we remove a ProcessRecord from the LRU
list when it may still have a process associated with it, report
that this happened, and try to make sure the process is killed.

Change-Id: I989b8c304f5fc14affb764c8535b13d6a1c7d89a