connectivity: Fix message in use error
NetworkMonitor makes use of a CustomIntentReceiver.
In the constructor of this it obtains a message using
two integers; it sends this message anytime it receives
an intent matching the action.  However, if two
intents are received in a short amount of time then it
risks sending the same message twice, therefore
the message in use error presents.  Moving the obtain
message call into the intent ensures that a fresh
message is obtained from the pool every time a send
happens.

Change-Id: Iaffabc3dfe85e22591430dd07066f66ba19296e7
CRs-Fixed: 803164