Fix an apk icon animation corruption issue
When several apk icons group together to a folder, the icons
scale down to small icons first. Dirty region computation only
respects the transformed small icons region. However, since the
swap behavior is PERSERVED, the contents of larger icon region
in previous frame are kept in the current frame, this region is
not covered by the dirty region, thus contents of previous frame
presents and it is corrupted. The fix is to include the in rectangle
to enlarge the dirty region

Change-Id: I546da255cf18b3c948e8491f3a513762d7dd4fea