Bluetooth: Fix background/direct connect race condition
There is a race condition in BluetoothDevice.connectGatt
API where a request to connect to a BLE device in
background (auto-connect flag is true) is issued as
direct connect. Direct connect times out after 30
seconds with error=133 (GATT_ERROR) since device is not
around. For such device that broadcast only when they
are used, we need background connect support

Change-Id: Ibdeb6e55d6906c9fcb56f2584eacc559c4b40823