Reset waitingVisible flag if activity is already visible.
During resumeTopActivity if the previous activity is finishing and
next activity is already visible then reset waitingVisible flag
for previous one so that it would not be added to waiting visible
list and destroyed when idle timeout is scheduled. This is done to
prevent unexpected timeouts between finish activity and destroy
activity.

CRs-Fixed: 827600
Change-Id: I94a551959a6b746ac4d10f614039c682b1d2997e