WindowManager: remove freezing window to fix UI freezing issue
In some corner case, some window is at exiting state, but can't be
removed. because the binder transact failed, and the state of window can't change
so that will still resizing window,This will result the windownmanger fall into freezing state.
The fix will stop resizing this window when encounter this case

Change-Id: I3e41aacadcaaea197dd74d8b93a7ed73ee441ff1
CRs-Fixed: 838815