SystemUI: Add auto sync tile into quick settings
Add a quick setting to control the auto sync configuration.
So that users could enable/disable auto sync with a single
click on the auto sync tile.

The quick setting tile is added. By default this feature is
not enabled. Add "sync" into quick_settings_tiles_default to
enable it if needed by any carrier.

Change-Id: I88a93b55d354796a90b2b36ad7a51a63476bc4d7
CRs-Fixed: 851542