Bluetooth: Fix to ignore the misplaced SRMP header during get operation
As per the GOEP 2.0 spec, the SRMP header shall be received in the
first GET response, and may be received in consecutive GET response
packets to cause the Client to continue its wait. However, once the
SRMP header is not received in a GET response, the SRMP header shall
not be received in another GET response for the duration of the
operation, hence ignore srmp wait header, if it's received in middle
of operation.

Change-Id: Ib126e13be06e85ac742afb1172cae91e5698c156
CRs-fixed: 885102