getParcelableArray* behaviour is sometimes unpredictable
in some case when the data in a savedInstance the Parcelable
arrays stored by the application are not unparcel'ed correctly
because the ClassLoader doesn't seem to be correct.

The first getSparseParcelableArray() in the code is preceded
by setClassLoader() by not the second one. The problem disappears
when setClassLoader() is called in both cases.

Fix same as:https://android-review.googlesource.com/#/c/40760/

Change-Id: I2a32d2fd1c628c9ef1b780dad73d4192176f7de8