SystemUI: Set Occluded true when keyguard is not showing
When entering incall view, "Occluded" was not set true in time, then it
would block KeyguardViewMediator.handleKeyguardDone() when screen locked.
Keyguard was still in showing state, so keyguardDone() and the lock sound
would be triggered many times.

Set Occluded true even when keyguard is not showing to update state asap.

Change-Id: I1d274044006db807c3d8be075a1c2665943f9fdb
CRs-Fixed: 892807