Fix for flicker when voice dialer activity launched
When a activity with windowFlag FLAG_SHOW_WHEN_LOCKED is launched
the background window ( in our case the launcher app window) to be
hidden in PhoneWindowManager.applyPostLayoutPolicyLw(), and to be
shown in WindowAnimator.updateWindowsLocked() continuously if there
is a show-when-locked dialog. This causes the wallpaper target to
change hence resulting in a flicker.

The check for keyguard locked bypasses the hideLw in
applyPostLayoutPolicyLw function and flicker is not observed.

Change-Id: If51f424523367cf73ad9fc0835e7ecd204db6144