perflock: Fix for boost duration during scroll.
During scrolls, to update duration of perflock, a check for timeout
to be 0 was present, but same timeout variable was updated after very
first scroll, with duration of scroll. Boost duration for successive
scrolls was not updating due to this check. Updated this conditional
logic to handle scroll boost duration properly.

Change-Id: I01449931215bfca0039335b3188e3101ce982d48