Fix HWUI texture leak
In bitmapShaderHelper, we try to get texture from texture cache.
If the texture is not in cache and can't fit in that cache, a
new texture is gererated. The generated texture is wrapped in
the BitmapShaderInfo* parameter and managed by AutoTexture when
bitmapShaderHelper returns. However, when the BitmapShaderInfo*
parameter is NULL, the generated texture is not managed and leak.
The fix is to use AutoTexture to manage it when shaderInfo is NULL.

Change-Id: Ieee133f112ff69e6e134129da10e0e2a36617f22