Softap: Display the count of connected devices to SAP
New status bar icon for wifi tethering is introduced which displays
the icon on the right corner. Tethering module displays a second
hotspot icon on the left corner in status bar when STA is connected
to SAP.
This commit removes the second hotspot icon from status bar and
display the count of multiple devices connected.

Change-Id: Ic27ba38fd1e30fddfb65ca6d7186f12d23b86e19
CRs-Fixed: 949267