MSIM: Fix to resolve SIM PIN unlock in APM.
In MSIM case, with two sims being locked on pin,
the second sim pin unlock screen does not get closed
even with the correct password.This is due to the
showDefaultMessage flag not being set to true after
sim pin is verified successfully.

Fix: In KeyGuardSimPinView, set the showDefaultMessage
flag to true after SIM PIN is verified successfully.

Change-Id: Ie31ee9fb009314555c81f3787d0ef3af4aad74c7
CRs-Fixed: 949288.