Fix the remaining attempts of PUK upon bootup.
If Sim is locked on PUK and is booted, currently user
is not informed about the remaining PUK attempts.In this
case, it can be possible that the sim becomes permanently
unusable on entering wrong PUK.

Fix: If Sim is locked on PUK, send the dummy empty string
as PUK password on bootup to get the available PUK attempts.
Also show the remaining PUK attempts to user on entering wrong
PUK password.

Change-Id: I90c7e01cdf4ce70a58b06c66fefe35bf76987fb8
CRs-Fixed: 955875