warm_boot: Define a variable to save SAP state
Warm Boot Requirement asks for SAP to be turned
off ( if ON ) and STA to be disconnected (if Connected)
on a Screen Off and vice versa on a Screen On case. To
accomplish this , the previous state of SAP state has to
be saved or restored. This commit defines a variable to
accomplish the same.

CRs-Fixed: 964598
Change-Id: Id6a05379c64ef55a1309410cdaa6d9d850997d63