Fix for activity not pausing on power key quick press
The top activity is not paused when the power key is pressed
quickly to suspend and resume device. The issue is due to device
entering WAKEFULNESS_DREAMING/DOZING state and lockscreen state
not being updated due to asynchronous design.

The Fix overrides the Dreaming behaviour to overcome this issue
by falling back to behaviour same as Dozing state.

Change-Id: Iae3d5b89094650b824a27d246d7ee0fc64cb787d