Add absent sim info to locked screen text
- For a multisim device, when only one sim is inserted
 that is pin locked, the carrier text only shows
'SIM CARD IS LOCKED - NO SERVICE' even though the
voice registration state for one of the phones is 12,
where emergency calls are allowed.
- This occurs when the state above is reported for the
slot where a sim card is not inserted. We only check for
active subscriptions when creating the display text
- Add a check for when only one sim card is inserted. If
the voice registration state is out of service for both
slots, with at least one as 12 (emergency allowed)
show 'Emergency calls only' instead of 'No service'.

Change-Id: Ib58b6b7431d7b29c770287037cbf07c8df101ae3
CRs-Fixed: 985955