DocumentsUI: Fix the IndexOutOfBoundsException in specific path
Check data cache state when using it

"State" referred in code fragment will be changed in some conditions.
But DocumentsUI has not setup the mechanism to response this change.
We can solve the problem by check and never import risk,
because the "State" is used in this code fragment only.

Change-Id: Ib9b57dc2a64b615325b14becdc8b81c405c1eea0
CRs-Fixed: 1003854
CRs-Fixed: 995407