Fix Emergency Button show issue
For multi-SIM phone, when there is only one SIM inserted
and it is in PIN state, the limited service may also be
reported from "NO SIM" slot. So it needs to check the
Service State of each slot to conclude if the phone could
make emergency call, other than just check the active
susbscription.

Change-Id: Ia78ee012effcb78a0a2f33e609f847265c372d38