Zygote: Additional whitelisting for legacy devices.
On M and below, we provide a blanket whitelist for
all files under "/vendor/zygote_whitelist". This path
is whitelisted purely to allow this patch to be applied
easily on legacy devices and configurations.

Note that this does not amount to a loosening of our security policy
because whitelisted files are reopened anyway.

Bug: 32691930
Test: manual
Change-Id: If5b53f6f0a707f8d36603c09bfd3f72dbfbbbb99