Temporarily permit reads in TelephonyManager.
Under strict mode, file reads/writes are not allowed in UI thread in order not to make UI less
responsive. A static member of TelephonyManager will be initialized in other application. E.g
Chrome. The TelephonyManager will be used as long as chromium calls getNetworkCountryIso().
Procfs is not a real filesystem and reading from it won't block (the kernel will just return
the information immediately), so it's not a problem to do it on the UI thread.

Bug: 120106498
Test: run cts -m CtsHostsideWebViewTests
              -t com.android.cts.webkit.WebViewHostSideStartupTest#testStrictMode

Change-Id: I3e75de215ec9c73264a565a0c2056ef67ee71d8d