Tigger aw8896 codec firmware loading in bootanimation
Firmware of aw8896 codec won't be loaded until the first music is going to
play after bootup. This will lead to a issue that the first time's playback
has slow response or even has no sound if the music is very short.
Trigger firmware loding in an early stage by playing a short clip of music
to solve that issue.

Change-Id: I94dac42d311b4a8a7c7f3a5e8359063881b7e502